% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishp.R
\name{fishp}
\alias{fishp}
\title{Calculate Fisher's method p-value and meta-analysis statistics}
\usage{
fishp(df, vars, df_sigma, sum_sigma)
}
\arguments{
\item{df}{data frame with "markname" and study names as column names.}

\item{vars}{character vector of study names to include in the meta-analysis.}

\item{df_sigma}{data frame of tetrachoric correlations.}

\item{sum_sigma}{sum of tetrachoric correlations.}
}
\value{
A data frame with columns 'markname', 'sum_chisq', 'sum_z',
'sum_sigma_var', 'pvalue', 'meta_z', 'meta_p', 'meta_nlog10p'
}
\description{
Calculate Fisher's method p-value and meta-analysis statistics
}
\examples{
  df <- data.frame(scan1 = generate_random_p_values(25, 0.1),
                   scan2 = generate_random_p_values(25, 0.1),
                   scan3 = generate_random_p_values(25, 0.1),
                   markname = LETTERS[1:25])
  varlist <- c("scan1", "scan2", "scan3")
  tc <- tetracorr(df, varlist)
  fishp(df, varlist, tc$sigma, tc$sum_sigma)
}
