% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polycorr.R
\name{polycorr}
\alias{polycorr}
\title{Calculate Polychoric Correlations}
\usage{
polycorr(data, varlist)
}
\arguments{
\item{data}{A data frame with variables to correlate}

\item{varlist}{A vector of variable names to correlate}
}
\value{
A data frame with polychoric correlation coefficients and standard errors
}
\description{
This function calculates the polychoric correlations between pairs of variables
in a given data frame. It returns a data frame with the row and column names of
the variables, the polychoric correlation coefficient, and its standard error.
}
\examples{
  data(mtcars)
  varlist <- c("mpg", "cyl", "hp", "drat", "wt", "qsec")
  polycorr(mtcars, varlist)

  data(mtcars)
  varlist <- c("mpg", "cyl", "hp", "drat", "wt", "qsec")
  polycorr(mtcars, varlist)
}
\seealso{
\code{\link{polychor}}
}
\author{
Woo Jung
}
\keyword{correlation,}
\keyword{polychoric}
