% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{z2p}
\alias{z2p}
\title{Convert z-score to p-value}
\usage{
z2p(data, out = NULL, pvalue = "pvalue", z, ztail = "P", type = "P")
}
\arguments{
\item{data}{A data frame or tibble.}

\item{out}{An optional character string indicating the name of the output data frame. If omitted, the output will have the same name as the input data frame.}

\item{pvalue}{An optional character string indicating the name of the output p-value column. Default is "pvalue".}

\item{z}{A numeric vector or scalar indicating the z-score(s).}

\item{ztail}{A character string indicating the tail of the distribution. Must be either "NORMAL" or "CHISQUARE". Default is "P".}

\item{type}{A character string indicating the type of p-value to calculate. Must be either "P" (parametric) or "N" (non-parametric). Default is "P".}
}
\value{
A data frame or tibble with the p-value column added.
}
\description{
This function calculates the p-value from a given z-score. The p-value can be parametric or non-parametric, and the tail of the distribution can be either normal or chi-squared.
}
\examples{
# Calculate the parametric p-value for a z-score of 1.96
z2p(data = iris, pvalue = "pvalue", z = 1.96, ztail = "NORMAL", type = "P")

# Calculate the non-parametric p-value for a z-score of -1.5
z2p(data = iris, pvalue = "pvalue", z = -1.5, ztail = "NORMAL", type = "N")

# Calculate the parametric p-value for a z-score of 2.5 and output the result to a data frame called "mydata"
mydata <- z2p(data = iris, out = "mydata", pvalue = "pvalue", z = 2.5, ztail = "NORMAL", type = "P")

}
