% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishp.R
\name{fishp}
\alias{fishp}
\title{Calculate Fisher's method p-value and meta-analysis statistics}
\usage{
fishp(df, vars, df_sigma, sum_sigma)
}
\arguments{
\item{df}{A data frame with columns 'markname' and study names as column names}

\item{vars}{A character vector of study names to include in the meta-analysis}

\item{df_sigma}{A data frame of tetrachoric correlations}

\item{sum_sigma}{Sum of tetrachoric correlations}
}
\value{
A data frame with columns 'markname', 'sum_chisq', 'sum_z',
'sum_sigma_var', 'pvalue', 'meta_z', 'meta_p', 'meta_nlog10p'
}
\description{
Calculate Fisher's method p-value and meta-analysis statistics
}
