% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{generate_random_p_values}
\alias{generate_random_p_values}
\title{Generates a list of random p-values with mixed significant and insignificant values}
\usage{
generate_random_p_values(n, ratio_significant)
}
\arguments{
\item{n}{number of samples}

\item{ratio_significant}{fraction of p-values to be significant}
}
\value{
list of n randomly generated significant and insignificant p-values
}
\description{
Generates a list of random p-values with mixed significant and insignificant values
}
