% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{pvalues_to_zscores}
\alias{pvalues_to_zscores}
\title{Convert P-values to Z-scores}
\usage{
pvalues_to_zscores(df_pvalues)
}
\arguments{
\item{df_pvalues}{A data frame containing p-values}
}
\value{
A data frame containing Z-scores
}
\description{
This function takes a data frame of p-values and converts them to Z-scores
using the quantile function for the standard normal distribution.
}
\examples{
  df <- data.frame(scan1 = runif(10),
                   scan2 = runif(10),
                   scan3 = runif(10),
                   markname = LETTERS[1:10])
  pvalues_to_zscores(df)

}
\seealso{
\code{\link{qnorm}}
}
\author{
Woo Jung
}
\keyword{Z-scores}
\keyword{p-values,}
