% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{pvalues_to_zscores}
\alias{pvalues_to_zscores}
\title{Convert P-values to Z-scores}
\usage{
pvalues_to_zscores(df_pvalues)
}
\arguments{
\item{df_pvalues}{data frame containing p-values}
}
\value{
data frame containing Z-scores
}
\description{
This function takes a data frame of p-values and converts them to Z-scores
using the quantile function for the standard normal distribution.
}
\examples{
  data(snp_example)
  head(snp_example)
  pvalues_to_zscores(snp_example)

}
\seealso{
\code{\link{qnorm}}
}
\author{
Woo Jung
}
