% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tetracorr.R
\name{tetracorr}
\alias{tetracorr}
\title{Calculate Tetrachoric Correlations}
\usage{
tetracorr(data, vars)
}
\arguments{
\item{data}{A data frame with variables to correlate}

\item{vars}{A vector of variable names to correlate}
}
\value{
A list containing a data frame with tetrachoric correlation coefficients,
and the sum of the tetrachoric correlations if the input variable was in
p-value form.
}
\description{
This function calculates the tetrachoric correlations between pairs of variables
in a given data frame. It returns a list containing a data frame with the tetrachoric
correlation coefficients, and the sum of the tetrachoric correlations if the
input variable was in p-value form.
}
\examples{
  data(mtcars)
  vars <- c("mpg", "cyl", "hp", "drat", "wt", "qsec")
  tetracorrs(mtcars, vars)

  data(mtcars)
  vars <- c("mpg", "cyl", "hp", "drat", "wt", "qsec")
  tetracorrs(mtcars, vars)
}
\seealso{
\code{\link{polychor}}
}
\author{
Woo Jung
}
\keyword{correlation,}
\keyword{tetrachoric}
